Specify tag for ubuntu image and apt update before install
The default latest tag was causing the image to fail to build due to not having sudo installed. An apt-get update is also needed to be run before the install otherwise the package is not found. A different mirror for hiawatha 9.13 also needs to be used as the previous path no longer hosts 9.x.